Improve delete all job performance
delete all jobs was deleting one job at a time using the jenkins
doDelete rest endpoint which was very slow.  Change Ia4fbfca970165
allows us to execute a groovy script on the jenkins server to
do the same thing.  The only difference is that running the script
on the server delete all jobs much faster.

As a data point:
I added all ~5,000 openstack jobs to my jenkins server.
The previous delete_all took many hours to delete all the
jobs while the script method did it in a matter of seconds.

Change-Id: Id7f7dfb567997e042fe0f783e54680482fefdc15
depends-on: Ia4fbfca970165d890d7e076f47ddcde7633afa9b