Update patch set 1
Patch Set 1: Code-Review-2

Indeed Wayne is quite correct, this is just one of a number of ways of specifying a string in yaml.

Sorry for the delay in reviewing, think that this should be abandoned as it appears to be due to confusion on available formatting specifiers available for strings in yaml.

Patch-set: 1
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2