Update patch set 4
Patch Set 4: Code-Review-1

(6 comments)

Looking for one more unit test, can probably also reduce the verbosity of the commit message. And a bug fix is needed for the Example in the docstring to ensure it renders correctly.

Haven't had to deal with contributions where the original Author hasn't been a party to the contributors agreement. I'm poking IRC to see if there is anything special required in such cases. Hopefully that won't take too long to clarify.

Patch-set: 4
Label: Code-Review=-1