Update patch set 18
Patch Set 18: Code-Review-1

(20 comments)

Wow, a lot has been added to this patch, thanks for the detailed doc strings as well :)

Sorry for taking so long to get back to reviewing it.

Some tidy up nits around the formatting, I think some of items where they are comma/space separated in the XML should be taking a list of items as the input, in some cases you already have the code doing close to the right thing but the docstring states the wrong argument type.

For the examples I think they just need to set more of the non default values to give more complete examples for people to follow. Particularly important around where reading in lists of data.

Patch-set: 18
Label: Code-Review=-1