Update patch set 4
Patch Set 4: Code-Review-1

(5 comments)

Few more changes and I think we should be good to land.

Should suggested structure changes based on improvements made previously to switch from using preformatted comma separated strings to using lists and converting in the code. Also I think an improvement to tweak the input format for the select-days to make it easier to spot and incorrect input and raise an exception for the user.

Patch-set: 4
Label: Code-Review=-1