Update patch set 6
Patch Set 6: Code-Review-1

(2 comments)

I havent carefully reviewed the code, but tests looks good to me.

An nit/nice to have: in the generated doc, the list of parameters for build-cause and day-of-week are rather hard to read:

http://docs-draft.openstack.org/90/189590/6/check/gate-jenkins-job-builder-docs/cba9cc4//doc/build/html/builders.html#builders.conditional-step

Can you make them a list in the generated Sphinx doc?

Patch-set: 6
Label: Code-Review=-1