Update patch set 2
Patch Set 2:

I added a change to do a list for another plugin here:

    https://review.openstack.org/196377/

As Darragh mentioned in comment in that patch, many plugins are moving support from comma to a list format so a helper function would be a good idea. If a helper function was created I'd be happy to rebase my change to use it.

Patch-set: 2
Reviewer: Gerrit User 13807 <13807@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0