Adds view functionality
Gives Jenkins Job Builder the ability to work with views using new methods in
python-jenkins to work with views.
Views are able to be created, updated, and deleted.
New modules for List view and Build Pipeline view are added.
Entry points are created for both respectively in a new section for views.
Example Yaml and Xml files are also included for each module in folders for
testing both build pipeline and list views.
There are also new tests for testing the deletion of views.

Change-Id: Idb29a4407bcc14593e10a4d951036cb04e8e6c27