Update patch set 12
Patch Set 12:

(4 comments)

I think my suggestion to _not_ special case the 'use-previous-build-as-reference' param in build_trends_publisher method is valid.  I also believe that what Dawid has done is pretty safe and that jenkins will ignore any xml it doesn't understand (my comments from PS 9).  

However now that this patch puts 'use-previous-build-as-reference' param is in the common build_trends_publisher method i'm not sure whether the modules that use that method all need to document that param.  The param could be missing for a few different reasons: author of the module forgot to add it to the doc, it's newly supported, or that it isn't supported.

I guess the only way to know is to verify that the plugin either supports that param or not.  I've identified where the params are missing inline.

Patch-set: 12
Reviewer: Gerrit User 6987 <6987@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0