Update patch set 5
Patch Set 5:

I haven't groked this entirely but do have a few questions from a quick glance.

It looks like you've set a global param for this feature yet it only applies to one module (trigger_parameterized_builds).  I think this makes it difficult to understand the context of the param just from reading the docs.  Users may expect that the param should apply to all modules/plugins or at least to all plugins where jenkins is aware of parameter ordering.  That's my expectation since it's a global param.

I could be wrong, but I think your goal with this param is to share across multiple modules however I'm not sure that it can be applied globally.   If your intention is to apply this feature more globally it would be nice to provide some set of helper methods to allow other modules to take advantage of this feature.  I'm not sure how plausible that would be though.

Patch-set: 5
Reviewer: Gerrit User 6987 <6987@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0