Update patch set 5
Patch Set 5:

I believe we have a number of places where we are not taking the ordering from yaml and this is just the first one. As we fix the remainder, it seems reasonable to hide this behaviour until we're ready to do a breaking change release (which I think should be after wayne's API rework).

I don't know if we can create helpers that will work in many situations, without defining a standard format for a mapping dict that can then be used by most modules to handle the conversion?

That might be a little more that should be tackled for fixing this.

Once we've fixed all places affected should result in being able to remove PYTHONHASHSEED=0 from the tox.ini and have it consistently produce identical XML on python 3.

Patch-set: 5
Label: Verified=0