Update patch set 1
Patch Set 1:

It might require a uniform indent on the section of code, or if you're running 'vim' typing 'gg=G'.

I suspect that the main reason we had formatting issues was people running 'gg=G' without 'shiftwidth=2' and 'expandtab' set to match how the Jenkins XML was originally formatted.

Looking at the original Jenkins files copy'n'paste would correctly be indent by 2 spaces and no use of tabs.

The main concern about not treating whitespace as significant is we had a few areas where we were testing content inside of tags where the whitespace was different between what was tested and what would really be outputted and thus uploading to Jenkins. There doesn't appear to be a way with doctest to distinguish between significant whitespace i.e.:

  <sometag> what is inside here
    should retain it's existing whitespace

    as it is considered significant by XML
  </sometag>

Instead we would test that:

    <sometag> is this the same</sometag>

The same as:

    <sometag>is
  this
  the
  same</sometag>

Which as far as XML is concerned, they are not, and given the use of particular characters as text separators in Jenkins, it is probably a bad idea to continue to treat them as the same.

Patch-set: 1