Update patch set 3
Patch Set 3:

Problem encountered before was due to the list of modified jobs being returned was blank resulting in the call to remove old unmanaged jobs wiping the server!

We now have tests to ensure this doesn't happen, so it should only need a very minimal check to ensure nothing else has been missed when running on the command line. I'll try to confirm it myself, already happy with the code with the new tests in place.

Patch-set: 3