Update patch set 1
Patch Set 1:

(1 comment)

Just wondering on the status of this change? We have exactly the use case requirements that this is intended to address. (We've been using a bit of a hack locally to do something similar but it's hard to manage whilst keeping up-to-date with upstream)

Patch-set: 1
Reviewer: Gerrit User 7015 <7015@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0