Update patch set 1
Patch Set 1:

(4 comments)

Code looks good to me, just think adjust the docstring format will make the source slightly easier to read.

One question on some of the exceptions, as you're already familiar with the MissingAttributeError just looking for clarification as why you haven't used it.

Perhaps we should add a new one in this case for EmptyAttributeError and add a helper function to raise either than or MissingAttributeError depending on the value? Could be done as a separate change, so only really think a tweak on the docstring needed here.

Patch-set: 1