Update patch set 3
Patch Set 3: Code-Review-1

I haven't reviewed yet but just wanted to point out there seems to be some documentation that can be added for this change.  Could you please take a look at the doc/source folder to see where it would be appropriate to add docs?  For example I think we need to add a section to installation.rst for this feature and would be good to create a view_list.rst and view_pipeline.rst files to expose the docs and examples from those files.

Patch-set: 3
Label: Code-Review=-1