Update patch set 1
Patch Set 1: Code-Review-1

Previously Sphinx is pinned < 1.2 and the setup.cfg has:

 [build_sphinx]
 all_files = 1
 build-dir = doc/build
 source-dir = doc/source
 builders = html,man,linkcheck

Using sphinx-build directly duplicate some of the information there and skips the man and linkcheck builders.

I would rather keep the build_sphinx but with this patch pbr raise an exception due to a method signature change:

 $ python setup.py build_sphinx
 running build_sphinx
 [pbr] Writing ChangeLog
 [pbr] Generating ChangeLog
 [pbr] Generating AUTHORS
 Traceback (most recent call last):
   File "setup.py", line 19, in <module>
     pbr=True)
   File "distutils/core.py", line 151, in setup
     dist.run_commands()
   File "distutils/dist.py", line 953, in run_commands
     self.run_command(cmd)
   File "distutils/dist.py", line 972, in run_command
     cmd_obj.run()
   File ".tox/docs/lib/python2.7/site-packages/pbr/builddoc.py", line 189, in run
     self._sphinx_run()
   File ".tox/docs/lib/python2.7/site-packages/pbr/builddoc.py", line 133, in _sphinx_run
     sphinx_config.init_values()
 TypeError: init_values() takes exactly 2 arguments (1 given)


Seems we might want to wait for PBR to catch up.


Is read the doc running Sphinx 1.3.1 regardless of what we have pinned in our requirements?

Patch-set: 1
Label: Code-Review=-1