Update patch set 1
Patch Set 1: Code-Review-1 Workflow-1

Ahh, so there's where the sphinx command lives.  i was trying to look for it in tox.ini.  Anyways to answer your questions, yes, i've been told by a core (agj) in #readthedocs channel that RTD currently only supports builds with ver 1.3.1.  You cannot force it to build docs with a different version of sphinx.

It looks like there was some work to fix this in pbr but it got reverted: I343c26560bfe5116d5

Not sure when that's gonna get fixed but i would agree that we should wait for pbr.

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=-1