Update patch set 5
Patch Set 5:

So more to add to doc/source/installation.rst?

I'm not ecstatic either about needing to work around things in this way, but I think it alleviates some problems in being able to land fixes to behaviour on the way to V3 without resulting in the API rework for V2 being impacted by changing the output generated.

Regarding identifying that the option is global in nature and shared, where best to add that? as a warning if enabled? simply noting the following along side the option doc:

'Enabling any __future__ option may result in changes in behaviour of multiple modules occurring at the same time as these options are global in nature and not limited to individual modules'

Patch-set: 5