Update patch set 5
Patch Set 5:

One of them has a 'name' attribute the other doesn't is the only difference. This also checks that whether the name is present or missing it doesn't get as far as throwing a KeyError exception on trying to access 'name'.

AFAICT this is intended to catch where the yaml is valid, but it doesn't follow the expected layout. Instead of throwing a generic exception it tries to classify it and include additional information to help the end user.

We weren't looking for yaml exception errors before and this isn't either. Adding a check to catch exceptions from yaml and also provide additional information would be a nice add on and seems like it could be built on top of this exception.

Right now, this is probably the minimal amount of test yaml required to exercise this (i.e. valid yaml but wrong structure for what was expected), so is there something we want to demonstrate by expanding it with additional yaml that won't get read?

Patch-set: 5