Update patch set 2
Patch Set 2:

IMHO, this is not good idea because plugin can't be customized for  particular job and zulip is a private chat.
Only one zulip account can be used across all gating jobs. And all people who wants use zulip notifications should have access to this account. But as I know it is not doable.

But core reviewers can have another vision to this patchset - You can ask them in IRC directly.

Patch-set: 2