Update patch set 3
Patch Set 3:

(2 comments)

Feedback should be addressed now for both builders and publishers.  I pulled most of the code out into helper functions for code reuse.  I also changed it so in builders you can just list the stacks but in publishers you have to specify whether it is create-stacks or delete-stacks.

Patch-set: 3