Update patch set 21
Patch Set 21: Code-Review-1

(10 comments)

Few minor nits around docstrings, examples and default values.

I'm a little concerned that in trying to flatten the settings namespace it makes it more obscure as to which options are related and interdependent. While the plugin may safely ignore certain options when other pre-requisites are not set, this does seem likely to cause confusion for users as to why options are being ignored when a job gets uploaded to Jenkins.

I'll let another core chime in on whether we need to push for a bit more structure/hierarchy before fixing the settings in place as adding code in the future that is backwards compatible tends to be much more onerous than adding modules.

Patch-set: 21
Label: Code-Review=-1