Update patch set 10
Patch Set 10:

Yes, i think you misunderstood my comment about complete001.yaml.   My suggestion was to create a new test that's similar to complete001.yaml.  You'll notice that most of the snippet examples we use for testing do not have '- job', '- project', or '- template' in them.  Those only run in our test harness but do not work when you pass them to the 'jenkins-jobs' command.  I think we need a test for views that a user can pick up and run as is.

Patch-set: 10
Label: Verified=0