Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

I'll let another core chime in on whether the changing what is assumed to be the default plugin version is a good idea, but generally it's better not to deviate from previous behaviour. And since it's possible to provide a config file containing plugins and versions to enable the new behaviour as well as pointing at the working jenkins instance will retrieve the correct plugin version for test runs (i.e. generating the XML output without uploading) it seems like it's unnecessary to change arbitrarily.

Also I would note that there is a change to have hipchat from 1.9 supported as a standard publisher via Ifa5e147da77882390d07e1bc9b58abff6f15ac1e so I'd be inclined to rebase on top of that (I'll update it to the latest) and call dispatch to use the function from the publisher module instead if the version is >= 1.9, with a deprecation warning logged as well to suggest users switch their config to use the publisher approach.

Patch-set: 2
Label: Code-Review=-1