Update patch set 12
Patch Set 12:

Moved configuration stuff from BaseTestCase into a callback and added a derived class that ensures a ConfigParser instance is returned to sanely test the slack integration. Fixed contentwise problems regarding the fact that "get()" was called but it was the implementation of pythons dictionary.

Furthermore cleaned some code inside the slack integration.

Patch-set: 12