WIP Fix defaults application.
Prior to this change defaults were applied differently when set for job-template
and job-group/project. This commit unifies the behavior such that both JJB
variables and JJB expansion parameters (triggers, builders, etc) will be pulled
in from defaults set at both places (job-template vs job-group/project).

It also adds the ability to specify a list of defaults that successively merge
on top of each other. This commit is WIP because it needs tests to validate this
behavior.

We also remove redundant applyDefaults call for non-template jobs.

Change-Id: I5e42815754790bc6f62f5b5dcc7260864c9cf447