Update patch set 2
Patch Set 2:

So Jenkins provides a plugin called the credentials plugin.
 https://wiki.jenkins-ci.org/display/JENKINS/Credentials+Plugin

This plugin provides a secure way to manage passwords in Jenkins. My question was does the Jenkins stash publisher plugin not use this feature of Jenkins?

If it does not that's really unfortunate and I would try to push for the developers of that plugin to support the credentials plugin but that's outside the scope of this patch. Hope this helps.

Patch-set: 2
Reviewer: Gerrit User 13807 <13807@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0