Update patch set 2
Patch Set 2:

Ok, got your point, but it doesn't look like it is using the credentials plugin. Now, even if it would, a change in JJB will still be needed to be able to set it up.
Appart of that, is it something else which looks wrong in my proposition ?

Patch-set: 2
Label: Verified=0