Add expansion check for shell-builder with no params
If you consider the following

- builder:
    name: test_builder
    builders:
      - shell: |
           echo ${{PASS_THROUGH}}
           echo ${FAIL}

- job-template:
    name: '{foo}-test'
    builders:
      - test_builder
      - shell: |
           echo ${{PASS_THROUGH_TOO}}

- project:
    name: 'foo'
    jobs:
      - '{foo}-test':
         foo: bar

You would expect this to fail with a message about "{FAIL}" not being
expanded.

As it stands, the expansion never happens for "test_builder" if no
parameter is passed in.  This leads to an inconsistency where if you
have a builder with no parameters, "${BAR}" gets passed through
verbaim as correct shell code, but if you add a parameter to that
builder, you now need to write "${{BAR}}" to get "${BAR}" in the
output.

The job-template builder has already been expanded, so we don't want
to re-expand this.  Hence the checks for the OrderedDict returned by
the YAML parser.

Change-Id: I9c0091cf08e87cb95d6c6e40c53c3bc59d2811a8