Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

So don't really know stash or how it even works in jenkins just commenting on your implementation.

So i'm wondering if stash can authenticate with a Jenkins token?  If so..
1.  do you still need this? 
2.  can you read a token in from the jenkins config instead of password? 
  [stash]
    authtoken = foo

It looks like this is what the hipchat plugin does.

Also I think you need to add some documentation for this.  Would like to see this param added to the jenkins_jobs.ini-sample file and add a stash section to the execution.rst file.

Patch-set: 2
Label: Code-Review=-1