Update patch set 22
Patch Set 22: Code-Review-1

(2 comments)

Sorry, looks like just one small thing around a defaults and the deployment-exclude-patterns was missed.

I'm presuming that the best solution is to allow the defaults for the exclude_patterns to be overridden so that for the deployment one, you can pass an empty list instead.

Patch-set: 22
Label: Code-Review=-1