Update patch set 7
Patch Set 7:

Guess I should have been included the use case that lies outside of this for plugin configuration. 

What do you name as the section for a plugin that can be configured via different modules? Ask users to specify two sections, one for each?

Or go with a hierarchy, with the more specific <module>_<plugin> allowed to override the generic if such a situation arises?

The cloudformation one comes to might here.

Patch-set: 7