Update patch set 9
Patch Set 9: Code-Review-1

(3 comments)

I didn't realize that hipchat publisher is a replacement for hipchat notify (implemented in hipchat_notify.py).  Would be good to consolidate these modules in a follow on patch.  Also hipchat notify is setup to read  a few params from the jjb config file which would be nice to add to this module but again can be done in subsequent change.   There's also a bunch of existing tests in tests/hipchat/fixtures that can be reused to validate the functionality of this publisher.

Patch-set: 9
Label: Code-Review=-1