Update patch set 2
Patch Set 2:

I still see the need of skipping "server" if that is empty. For the general usage, people won't be providing that server, and will expect to use the defaults for the jobs.
If you force it, you will end up with people having empty <server></server> in their xml.
I think it's easy to add a comparison for "if server" and only populate in case it has a value set.

Also, different question. I see that now this patch expects that server is populated on the publisher. Does it mean that we have to setup server var on each job we define? I'd like to have a way, either by querying server var configured in the plugin, or having a var in jjb ini file, to setup the server globally. And we could have this patch as well, that will allow us to specify a server per job.

Patch-set: 2