Update patch set 2
Patch Set 2:

It defaults to api.hipchat.com if no server is provided.  What I'll do is change this to see if server was defined in jjb and if it was, I'll include the server block to override the default, otherwise, the default should be used which is set in the plugin.

Patch-set: 2