Update patch set 3
Patch Set 3: Code-Review-1

While testing this I found an issue with delete-old option:
  https://storyboard.openstack.org/#!/story/2000357

Which makes me wonder if this feature is meant to fix the 'delete-old' option or is it an addendum to it?  I mean is 'delete-unmanaged' is how 'delete-old' was supposed to work?

Patch-set: 3
Reviewer: Gerrit User 6987 <6987@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1