Update patch set 3
Patch Set 3:

Ok. thanks for the explanation.  I've updated the bug with a possible fix for delete-old but since that bug also applies to delete-unmanaged I would be hesitant to push this thru without fixing that bug first.  

Another thing i was wondering whether this would support a third option?
  1. delete only jobs that should be managed by JJB and leave manually created jobs as is 
  2. delete all jobs if your server is fully managed by job
  3. delete all unmanaged jobs while keeping manage jobs (inverse of 1)?

Patch-set: 3