Use xml instead of job description for is_managed
Inject a jjb xml attribute into the job to determine if a job is
managed. Old method relied on modifying the description with a magic
string which followed a job on copy. The xml attribute does not follow
the job on copy making for a more reliable determination of management.

XML tree follows the following format:
<jjb>
    <managed>true</managed>
</jjb>

story: 2000357
Change-Id: I318a5901052fe54664c8578b51c07a1aa01b8586