Update patch set 3
Patch Set 3:

I suspect that option 1 would be useful only if you intend to have a self-destruct button/someone decides that JJB isn't right for them. I'm game to fix it if you are.

For now I've opened a new branch/review for it since that is more a bug in the "is_managed" function and not so much the delete-{old,unmanaged}. Also, that change is "potentially dangerous" as it's going to cause a mass update the first time it runs!

is_managed changes here:
https://review.openstack.org/#/c/225562/

Patch-set: 3