Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Seems like this would fix most of the concerns.

The only gaps remaining are:
# running of a delete operation before doing an upload with the new tag will not do the correct thing on the first release.
# if someone manually saves a JJB managed job that is subsequently deleted from the JJB config, I'm presuming the manual save will remove the extra XML attributes and it won't get removed on the new clean-up if only old managed jobs are being removed.

I think perhaps for one release we need to allow jobs be regarded as managed if either the MAGIC string or the new attribute appear. Output a warning message about the upcoming change in behaviour and then for the next release remove support for the old magic string check.

Definite discussion point ^^

Patch-set: 1
Label: Code-Review=-1