Update patch set 2
Patch Set 2:

Just re-read the code, the problem isn't that we're not checking for the default configuration, it's that we leave 'options.conf' set to the default of None even when we do discover a valid configuration.

I think we should change the behaviour there to ensure that options.conf reflects the conf file used for configuration. This allows detection of whether internal only configuration is enabled or a configuration file was made available via option or from the default system wide location.

That way, we can make the use case to perform the plugins lookup if we have a valid config file, whether it's user or system default.

I think set options.conf to the system default if it exists and the user didn't provide a custom on.

Patch-set: 2