Update patch set 4
Patch Set 4: Code-Review+1

Rebased this to the latest master and modified a test to pass on a more recent (and strict) version of mock.

I've exercised this against a local jenkins instance to see that changes to some jobs are correctly updated in parallel without accidentally removing other jobs which was the reason for a revert previously.

Uploading all ~7k jobs to a jenkins instance could be done in < 4minutes using 50 workers. I suspect that openstack should be able to enable some parallel workers now given that a previous fix ensured we would cache the list of jobs that were on the server for the entire run instead of checking once for each upload.

Doing some monitoring using jconsole that for 50 workers it used about ~2GB memory from the heap space temporarily, so provided you have >2GB free available in the heap you could probably go higher.

Patch-set: 4
Label: Code-Review=+1