Update patch set 4
Patch Set 4: -Code-Review

I can say the delete-all/rebuild is undesirable for my use case specifically because of what Darragh Bailey said.

As an "end user" of JJB I plan on running something akin to delete-unmanaged every 5 minutes to be sure that no one sneaks a job into jenkins without putting it into version control. I envision it more as an enforcement tool and less as a cleanup!

Patch-set: 4