Update patch set 4
Patch Set 4:

(1 comment)

I am generally in favor of this feature, in fact I have been working recently on python scripts using python-jenkins to handle a similar use case. During discussion with my team about this topic here are the types of unmanaged jobs I noted (ci-job-configs is the git repo we use to manage JJB configs):

1 Job was previously managed by ci-job-configs,
1a Became unmanaged because it is no longer needed (so it must be deleted).
1b Became unmanaged because it is being replaced by a job with a
different name and a different behavior.
1c Became unmanaged because of a change of name (job does the same
thing as previously, but for some reason the job template was renamed
leading to duplicates of the same job). Depending on whether or not
there was a job type change involved, this may be as simple as
renaming the existing job just prior to merge/deploy of new job.
2 Job was deployed as an experimental copy for the sake of pre-merge
testing. (we prefix such jobs with the string "experimental_" so it
is always possible to distinguish these types of jobs form one another).
3 Job was manually created rather than created by JJB.

The tooling I am working on around these various types of unmanaged jobs is intended to allow us to easily distinguish between (1), (2), and (3) for the sake of automated deletion of (2) and (3). It's a little more difficult to automatically distinguish between (1a), (1b), and (1c) so our team is also working on a formal process to ensure that we notice these types of unmanaged jobs and take the appropriate steps to prevent them from being deleted as if they were (3).

Preventing jobs from being deleted in the case of (1c) may be important because deleting jobs leads to loss of history--important for us because some of our product dev teams rely on that history while it's available.

Anyway, hope that's helpful for discussion.

I'm not against the extra unused parameter in the new Builder method by the way, I'm just wondering if anyone involved here actually uses JJB as anything other than a command line tool. If not, then it appears that we are developing and testing against features that no one currently uses? Don't get me wrong, it looks hypothetically useful--I'm simply curious.

Patch-set: 4
Reviewer: Gerrit User 13487 <13487@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0