allow modules to manipulate data structure per job
The motivating use case is, for configuration that takes the form of a
list, to provide a mechanism to conditionally insert a value.
Previously a variable like `{foo}` could have the empty string or
`'None'` (both of which may have meaning to plugins) but not be
omitted.  For example, for multiple projects in the same repository
with a fixed set of different dependencies, it would be useful to
configure git sparse checkout paths and polling per project and *omit*
directors from polling.

`deep_format` leverages python's existing string formatting machinery
that (reasonably) always returns strings (such as `'None'`).  To allow
`None` to be passed through post-template a special case needed to be
added for templates with a single field whose value is `None`.  No
existing tests demonstrated reliance on the old string-None behavior.

Recursive filtering code is adapted from examples in
http://stackoverflow.com/questions/20558699/python-how-recursively-remove-none-values-from-a-nested-data-structure-lists-a

Initially only enabled for the `scm` module.  If this works out I
expect support to be added to other modules.

Alternative approaches considered:
 * Add `[x for x in lst if x is not None]` for every use of a list in
   a module.  This quickly proved ugly and is obviously not general.
 * Use `handle_data`.  This was unworkable because `handle_data`
   operates on the entire configuration *before* templating occurs.
 * Do this in `gen_xml` of the module.  This would still require the
   formatting changes and I think it would be surprising for `gen_xml`
   to manipulate the data it is working on.
 * Treat `'None`' as the special value.  Using a value that is not
   falsy and that could be a legitimate variable for a job would
   likely break something.

Change-Id: I7d2323649819ae7eff4dd0dc1605b5287060b355