Update patch set 5
Patch Set 5: Code-Review-1

(9 comments)

Little bit confusing as to whether this is the plugin that is listed in the jenkins wiki or if it's something else provided by openshift. Would be good to ensure that any documentation linked to actually describes the plugin if possible instead of linking to the website entry page.

docstrings need some fixing to align with the preferred style, various links also need to be fixed/changed, avoid using '/' for line continuation and update the tests to have separate ones that verify the defaults (with minimal or no user settings) to ones that verify when the user overrides all settings.

Appears to be some confusion around how to create links in that it's half coping a custom macro used to simplify generation of links for the jenkins wiki. Recommend that you look at the docs generated to check how the links get formatted.


The code seems good but I didn't look as closely as I did to the tests/examples and docstrings.

Patch-set: 5
Label: Code-Review=-1