Fix default expansion in macros
Default formatting for external macros which used syntax like:

- parameter:
    name: example
        parameters:
      - string:
              name: OS_USERNAME
              default: '{OS_USERNAME}'

has never automatically received variables set higher in the chain. The
only way to pass these variables to the macro was to do it explicitly
each time the macro was invoked, which means a lot of repetition.

This patch fixes the expansion of such defaults, making external macros
require less repetition. It also does not break any existing
job-templates which use the previously-required syntax, and if a
value is specified explicitly when using a macro, like:

  parameters:
    - example:
        OS_USERNAME: 'demo'

Then the explicit definition overrides anything from further out, as it
should.

Change-Id: I2014100f52fd9c4556e7b20bf6e885c9c650a50b