Update patch set 3
Patch Set 3:

To include some comments shared in IRC.

The existing default is to output all logging for all tests interspersed. This means you still get to see the log messages. The problem is, when a test fails, how do you associated which log messages occurred for that specific test.

This change makes use of logging fixtures, which will attach the log output in front of the failed test output so it is easy to see what log messages were outputted for that test.

This improves the usefulness of log messages for test failures, though it still requires that we add useful log messages.

Only remaining change is to tidy up the base class since the base class is really just for the scenarios and isn't the base class for the command line tests.

Patch-set: 3