Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

So there is a very subtle difference in the helper function, which is resulting in a change to the XML outputted. Looks like you just need to change the mappings to use None instead of '' to retain the existing test output.

I think that should be all that's required. :)

Patch-set: 1
Label: Code-Review=-1