Update patch set 1
Patch Set 1: -Code-Review

What I suggested was to make the utility function more generic, which looks great now, since we are able to emit empty elements if we need. :)

As a result, the previous uses of that function which assumed that a val of '' meant no element need to be updated to conform to the new style. So you'd just need to update the call sites to use None instead of '' for val.

Patch-set: 1