Update patch set 3
Patch Set 3: Code-Review+2

Established via the IRC channel that the empty tag for the openshift functions matches the plugin behaviour in Jenkins. I updated the commit message to explain so it was clear that with the refactorying that the mapping was not changed when the helper function was modified to be generic.

Patch-set: 3
Reviewer: Gerrit User 1054 <1054@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2