cmd: Add a fallback for jenkins_jobs.ini to the user's home directory
The system-global path of

    /etc/jenkins_jobs/jenkins_jobs.ini

is not applicable on Windows. To solve that, and as Linux users would
benefit from a per-user configuration file, too, introduce a fallback at

    ~/.config/jenkins_jobs/jenkins_jobs.ini

following the XDG Base Directory Specification. Try to use that per-user
file if no configuration file is given on the command line and none is
present in the script directory for development purposes.

A test for this new feature is added in the new "tests/config" directory.
For this to work, stop ignoring "config" in Git. If that rule was meant to
ignore "config" files inside the ".tox" directory it is not necessary
anyway as ".tox" is already ignored.

Change-Id: I1a22454999ee8ac60d6f22545850c33aa03bb29b