Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Take a look at one of the python console outputs for example [1]. The issue is your xml you provided does not appear to be formatted as the unit tests expect. Particularly I think you need to use 2 space indents instead of tabs so that when JJB generates the xml it matches.

[1] http://logs.openstack.org/81/236981/1/check/gate-jenkins-job-builder-python34/0860eb6/console.html

Patch-set: 1
Label: Code-Review=-1