cmd.py: tweak pluging information querying logic
We don't want to unconditionally disable plugin querying just because
--conf is not given. That would mean that the query_plugins_info field
in the default paths would always be ignored, even if set to True.

This patch tweaks the logic so that it's not whether --conf or not is
given that determines info querying, but whether the configuration came
from a file or not.

This way, the `test` command will still work out of the box, as commit
f68831b envisioned. But the query_plugins_info is now checked if the
user has a configuration file (whether it be a default one, or one
specified through --conf).

Change-Id: Id71ad207f0419f7cc52eca174dd859d4182d7250
Story: 2000387