Update patch set 3
Patch Set 3:

Dong Ma, I noticed this as well. I have to rename my .ini file before running tox. Not sure what would be the correct way to handle this, if it does need to be handled. Is there already a policy/established way to deal with tests that get confused with the real install?

Patch-set: 3
Label: Verified=0