Update patch set 1
Patch Set 1: Code-Review-1

A simple test with pip 1.5.4 suggests it will break as it cannot parse the line, currently using tox for testing will result in a much newer pip being installed that can work with it.

Since technically we no longer support python 2.6, we could just remove it and leave it to anyone who wants python 2.6 support to patch it back in?

In any case, seems like this needs a little more thought on what to do.

Patch-set: 1
Label: Code-Review=-1